requirements.txt
streamlit
pandas
openai
matplotlib
seaborn
plotly
python-dotenv